refactor test code for loans pallet
